Add ceph-pool-weight option for calculating pgs
Provide the weight option to the Ceph broker request API for requesting
the creation of a new Ceph storage pool. The weight is used to indicate
the percentage of the data that the pool is expected to consume. Each
environment may have slightly different needs based on the type of
workload so a config option labelled ceph-pool-weight is provided to
allow the operator to tune this value.

Closes-Bug: #1492742

Change-Id: I844353dc8b354751de1af5d30b6d512712d40a62