Add use-syslog configuration option
This config option is used in the ceph.conf template; add to
charm configuration options to allow syslog to be enabled or
disabled for ceph library usage.

The existing context configuration for the charm will just
pickup and use this configuration option, avoiding the current
behaviour of writing out 'none' to the ceph.conf file.

Change-Id: I40debfa5c8ee07999ed5e688e31d1c6ceffbea36
Closes-Bug: 1604575